% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidths.R
\name{bw_beta}
\alias{bw_beta}
\title{Bandwidth selection for the beta kernel estimator}
\usage{
bw_beta(udata)
}
\arguments{
\item{udata}{data.}
}
\value{
optimal bandwidth parameter.
}
\description{
The bandwidth is selected by minimizing the MISE using the Frank copula as
the reference family. The copula parameter is set by inversion of Kendall's
tau. See Nagler (2014) for details.
}
\details{
To speed things up, optimal bandwidths have been pre-calculated on a grid of
tau values.
}
\references{
Nagler, T. (2014).
Kernel Methods for Vine Copula Estimation.
Master's Thesis, Technische Universitaet Muenchen,
\url{https://mediatum.ub.tum.de/node?id=1231221}
}
