% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidths.R
\name{bw_tll}
\alias{bw_tll}
\title{Bandwidth selection for the transformation local likelihood estimator}
\usage{
bw_tll(udata, deg)
}
\arguments{
\item{udata}{data.}

\item{deg}{degree of the polynomial.}
}
\value{
optimal bandwidth matrix.
}
\description{
The bandwidth is selected by a rule of thumb similar to \code{\link{bw_t}}.
}
\details{
The formula is
\deqn{5  n^{-1 / (4q + 2)}  \hat{\Sigma}^{1/2},}
where \eqn{\hat{Sigma}} is empirical covariance matrix of the transformed
random vector and \eqn{q = 1} for \code{TLL1} and \eqn{q = 2} for
\code{TLL2}.
}
