% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidths.R
\name{bw_tt_plugin}
\alias{bw_tt_pcv}
\alias{bw_tt_plugin}
\title{Nearest-neighbor bandwidth selection for the tapered transformation estimator}
\usage{
bw_tt_plugin(udata, rho.add = TRUE)

bw_tt_pcv(udata, rho.add = T)
}
\arguments{
\item{udata}{data}

\item{rho.add}{logical; whether a rotation (correlation) parameter shall be
included.}
}
\value{
bandwidth matrix
}
\description{
The smoothing parameters are selected by the method of Wen and Wu (2015).
}
\author{
Kuangyu Wen
}
\references{
Wen, K. and Wu, X. (2015).
Transformation-Kernel Estimation of the Copula Density,
Working paper,
\url{http://agecon2.tamu.edu/people/faculty/wu-ximing/agecon2/public/copula.pdf}
}

