% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.kdecopula}
\alias{print.kdecopula}
\title{Print a kdecopula obejcts}
\usage{
\method{print}{kdecopula}(x, ...)
}
\arguments{
\item{x}{kdecopula object}

\item{...}{}
}
\value{
\code{invisible(x)}
}
\description{
Print a kdecopula obejcts
}

