% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.kdecopula}
\alias{summary.kdecopula}
\title{Summary of kdecopula object}
\usage{
\method{summary}{kdecopula}(object, ...)
}
\arguments{
\item{object}{kdecopula object}

\item{...}{}
}
\value{
\code{invisible(object)} after adding information (if missing).
}
\description{
Summary of kdecopula object
}

