% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.kdecopula.R
\name{plot.kdecopula}
\alias{plot.kdecopula}
\title{Plotting \code{kdecopula} objects}
\usage{
\method{plot}{kdecopula}(x, type = "contour", margins, size, ...)
}
\arguments{
\item{x}{\code{kdecopula} object.}

\item{type}{plot type; either \code{"contour"} or \code{"surface"} (partial
matching is activated) for a contour or perspective/surface plot
respectively.}

\item{margins}{either \code{"unif"} for the original copula density or
\code{"norm"} for the transformed density with standard normal margins
(partial matching is activated). Default is \code{"norm"} for \code{type =
"contour"}, and \code{"unif"} for \code{type = "surface"}.}

\item{size}{integer; the plot is based on values on a \eqn{size x size}
grid; default is 100 for \code{type = "contour"}, and 25 for \code{type =
"surface"}.}

\item{\dots}{optional arguments passed to \code{\link{contour}} or
\code{\link{wireframe}}.}
}
\description{
Produces contour or perspective plots for a \code{kdecopula} object.
}
\examples{
## load data and transform with empirical cdf
data(wdbc)
udat <- apply(wdbc[, -1], 2, function(x) rank(x)/(length(x)+1))

## estimation of copula density of variables 5 and 6
obj <- kdecop(udat[, 5:6])

## plots
plot(obj)  # (marginal normal) contour plot
plot(obj, margins = "unif")  # contour plot of actual copula density
plot(obj, type = "surf")  # surface plot of actual copula density
}
\author{
Thomas Nagler
}
\seealso{
\code{\link[kdecopula:kdecop]{kdecop}},
\code{\link[graphics:contour]{contour}},
\code{\link[lattice:wireframe]{wireframe}}
}
\keyword{Copula}
\keyword{density}
\keyword{estimation}
\keyword{plot}

